pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib>=5.0.0",
    "beanie>=2.0.1",
    "emails>=0.6",
    "fastapi>=0.128.0",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "uvicorn>=0.40.0",
]
